Create README.md
#Machine Learning Final Project

This repository provides a demo on how to conduct speech recognition on commonly used words with an accuracy of about 95%. The words chosen for this project were 'yes' and 'no', but other words datasets can be used in this demo. 
This demo, after creating the model, allows you to record and test audio files within the notebook.

Other word pairs like 'no' and 'go' were tested with this model but the accuracy dropped to about 86% for words that sounded very similar to one another. Although the accuracy was not as high as with 'yes' and 'no', the model was still able to correctly identify the word being spoken.